Modernize build: portable CMake, vendored Lua 5.0, cross-platform source fixes
Make the original Dev-C++/MinGW (Windows-only) YurOTS sources for Tibia 7.6
build and run on modern GCC/Clang/MSVC across Windows, Linux and macOS, with
runtime behaviour preserved 1:1. Verified building clean and running on macOS
(Apple clang, arm64): loads the Lua config, all XML data, items.otb and the
OTBM map, binds port 7171 and answers the OpenTibia status query.

Build system:
- CMakeLists.txt mirroring source/devcpp/Makefile.win (48 of its 49 entries;
  stale builtinaac.cpp dropped), feature flags 1:1, XML-only storage.
- cmake/BuildVendoredLua.cmake builds PUC Lua 5.0.2 statically (SHA256-pinned)
  from the bundled source/libraries/lua-5.0.2.tar.gz — keeps the engine's
  Lua 5.0 C API unchanged (no script-facing change). No GMP/MySQL/SQLite.

Behaviour-preserving source changes (each commented inline with "modernization:"):
- definitions.h/luascript.h: hand-rolled fixed-width typedefs -> <cstdint>.
- items.h: __gnu_cxx/stdext hash_map -> std::unordered_map.
- otsystem.h/tools.cpp: winsock.h -> winsock2.h; _ftime/ftime/timeb -> std::chrono.
- removed "extra qualification on member" defs (fileloader/item/luascript/spells).
- ltoa/_ultoa/_i64toa/_atoi64 -> snprintf/strtoll; abs -> labs.
- 64-bit pointer round-trip through Lua widened int -> intptr_t.
- _HOMEDIR_CONF_ guarded so __NO_HOMEDIR_CONF__ reads ./config.lua everywhere.

Pre-existing bugs fixed (invisible on the original 32-bit Windows build):
- 64-bit OTB/OTBM reading: on-disk 32-bit fields were read through `unsigned
  long` (8 bytes on LP64), so items.otb and every map failed to load. On-disk
  fields made fixed-width uint32_t (GET_ULONG, header read, flags_t,
  VERSIONINFO, OTBM_root_header).
- tools.cpp upchar(): CP1252 accent table corrupted to U+FFFD; reconstructed
  as the standard Latin-1/CP1252 mapping.

See ots/MODERNIZATION.md for the full account.

+++ ots/source/actions.cpp
@@ -454,7 +454,7 @@ _player(NULL)
 	else
 		fclose(in);
 	lua_dofile(luaState, scriptname.c_str());
-	this->setGlobalNumber("addressOfActionScript", (int)this);
+	this->setGlobalNumber("addressOfActionScript", (intptr_t)this); // modernization: (int) truncated the pointer on 64-bit
 	this->loaded = true;
 	this->registerFunctions();
 }
@@ -716,7 +716,7 @@ int ActionScript::registerFunctions()
 
 ActionScript* ActionScript::getActionScript(lua_State *L){
 	lua_getglobal(L, "addressOfActionScript");
-	int val = (int)internalGetNumber(L);
+	intptr_t val = (intptr_t)internalGetNumber(L); // modernization: pointer-sized round-trip (see setGlobalNumber)
 	ActionScript* myaction = (ActionScript*)val;
 	if(!myaction){
 		return 0;
